refactor(frontend): consolidate shared ui and remove legacy components
