--- frontend/src/components/AdminPasswordPrompt.tsx
@@ -1,3 +1,0 @@
-export { default } from "@/shared/admin-password";
-export { AdminPasswordPromptContainer } from "@/shared/admin-password";
-export type { AdminPasswordPromptProps } from "@/shared/admin-password";
