--- frontend/src/components/admin-password/AdminPasswordPromptContainer.tsx
@@ -1,2 +1,0 @@
-export { default } from "@/shared/containers/admin-password/AdminPasswordPromptContainer";
-export * from "@/shared/containers/admin-password/AdminPasswordPromptContainer";
