--- frontend/src/components/admin-password/AdminPasswordPromptIdle.tsx
@@ -1,1 +1,0 @@
-export * from "@/shared/components/admin-password/AdminPasswordPromptIdle";
