--- frontend/src/components/admin-password/AdminPasswordPromptIncorrect.tsx
@@ -1,1 +1,0 @@
-export * from "@/shared/components/admin-password/AdminPasswordPromptIncorrect";
