--- frontend/src/components/admin-password/AdminPasswordPromptLayout.tsx
@@ -1,1 +1,0 @@
-export * from "@/shared/components/admin-password/AdminPasswordPromptLayout";
