--- frontend/src/components/admin-password/AdminPasswordPromptMissing.tsx
@@ -1,1 +1,0 @@
-export * from "@/shared/components/admin-password/AdminPasswordPromptMissing";
