--- frontend/src/components/admin-password/constants.ts
@@ -1,1 +1,0 @@
-export * from "@/shared/components/admin-password/constants";
