--- frontend/src/components/admin-password/types.ts
@@ -1,1 +1,0 @@
-export * from "@/shared/components/admin-password/types";
