--- frontend/src/components/ui/ErrorState.tsx
@@ -1,1 +1,0 @@
-export { default } from "@/shared/components/error/ErrorState";
