--- frontend/src/components/ui/error/AlertError.tsx
@@ -1,1 +1,0 @@
-export { default } from "@/shared/components/error/AlertError";
