--- frontend/src/components/ui/error/CardError.tsx
@@ -1,1 +1,0 @@
-export { default } from "@/shared/components/error/CardError";
