--- frontend/src/components/ui/error/CenteredError.tsx
@@ -1,1 +1,0 @@
-export { default } from "@/shared/components/error/CenteredError";
