+++ frontend/src/features/leaderboard/components/LeaderboardError.tsx
@@ -1,5 +1,5 @@
 import { Alert, Box } from "@chakra-ui/react";
-import CenteredContainer from "@/components/ui/CenteredContainer";
+import { CenteredContainer } from "@/shared/ui";
 
 interface Props {
   message: string | null;
