+++ frontend/src/features/tournament/components/TournamentError.tsx
@@ -1,5 +1,5 @@
 import { Button, Center, Heading, Text, VStack } from "@chakra-ui/react";
-import CenteredContainer from "@/components/ui/CenteredContainer";
+import { CenteredContainer } from "@/shared/ui";
 
 interface TournamentErrorProps {
   message: string | null;
