+++ frontend/src/features/tournament/components/TournamentNotFound.tsx
@@ -1,5 +1,5 @@
 import { Center, Text } from "@chakra-ui/react";
-import CenteredContainer from "@/components/ui/CenteredContainer";
+import { CenteredContainer } from "@/shared/ui";
 
 export default function TournamentNotFound() {
   return (
