+++ frontend/src/features/tournament/components/TournamentsView.tsx
@@ -1,5 +1,5 @@
 import { Box, Button, VStack } from "@chakra-ui/react";
-import CenteredContainer from "@/components/ui/CenteredContainer";
+import { CenteredContainer } from "@/shared/ui";
 import type {
   CreateTournament,
   Tournament,
