+++ frontend/src/main.tsx
@@ -1,13 +1,13 @@
-import { Provider } from "@/components/ui/provider"
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import './index.css'
-import App from './App.tsx'
+import { Provider } from "@/shared/ui";
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+import "./index.css";
+import App from "./App.tsx";
 
-createRoot(document.getElementById('root')!).render(
+createRoot(document.getElementById("root")!).render(
   <StrictMode>
     <Provider>
       <App />
     </Provider>
-  </StrictMode>,
-)
+  </StrictMode>
+);
