+++ frontend/src/shared/admin-password/components/AdminPasswordPromptIdle.tsx
