+++ frontend/src/shared/admin-password/components/AdminPasswordPromptLayout.tsx
