+++ frontend/src/shared/admin-password/components/AdminPasswordPromptMissing.tsx
