+++ frontend/src/shared/admin-password/components/index.ts
