+++ frontend/src/shared/admin-password/index.ts
@@ -1,7 +1,3 @@
-import AdminPasswordPromptContainer from "../containers/admin-password/AdminPasswordPromptContainer";
-
-export * from "../components/admin-password";
-export { default as AdminPasswordPromptContainer } from "../containers/admin-password/AdminPasswordPromptContainer";
-export * from "../containers/admin-password/AdminPasswordPromptContainer";
-
-export default AdminPasswordPromptContainer;
+export * from "./components";
+export * from "./containers";
+export { AdminPasswordPromptContainer as default } from "./containers";
