--- frontend/src/shared/components/index.ts
@@ -1,2 +1,0 @@
-export * from "./admin-password";
-export * from "./error";
