--- frontend/src/shared/containers/index.ts
@@ -1,1 +1,0 @@
-export * from "./admin-password";
