+++ frontend/src/shared/error/components/AlertError.tsx
@@ -1,6 +1,6 @@
 import { Alert, Box } from "@chakra-ui/react";
 
-import CenteredContainer from "@/components/ui/CenteredContainer";
+import { CenteredContainer } from "@/shared/ui";
 
 interface AlertErrorProps {
   title: string;
