+++ frontend/src/shared/error/components/CenteredError.tsx
@@ -1,35 +1,31 @@
-import { Box, Button, Text, VStack } from "@chakra-ui/react";
+import { Button, Center, Heading, Text, VStack } from "@chakra-ui/react";
 
-import CenteredContainer from "@/components/ui/CenteredContainer";
+import { CenteredContainer } from "@/shared/ui";
 
-interface CardErrorProps {
+interface CenteredErrorProps {
   title: string;
   message?: string | null;
   actionLabel?: string;
   onAction?: () => void;
+  height?: string | number;
 }
 
-export default function CardError({
+export default function CenteredError({
   title,
   message,
   actionLabel,
   onAction,
-}: CardErrorProps) {
+  height = "300px",
+}: CenteredErrorProps) {
   return (
     <CenteredContainer>
-      <Box
-        w="100%"
-        p={6}
-        borderWidth="1px"
-        borderRadius="lg"
-        borderColor="red.300"
-      >
-        <VStack gap={2}>
-          <Text color="red.500" fontWeight="semibold">
+      <Center h={height}>
+        <VStack gap={3}>
+          <Heading size="md" color="red.300" textAlign="center">
             {title}
-          </Text>
+          </Heading>
           {message && (
-            <Text color="red.400" fontSize="sm" textAlign="center">
+            <Text color="red.200" textAlign="center">
               {message}
             </Text>
           )}
@@ -44,7 +40,7 @@ export default function CardError({
             </Button>
           )}
         </VStack>
-      </Box>
+      </Center>
     </CenteredContainer>
   );
 }
