+++ frontend/src/shared/error/components/ErrorState.tsx
