+++ frontend/src/shared/error/components/index.ts
