+++ frontend/src/shared/error/index.ts
@@ -1,0 +1,1 @@
+export * from "./components";
