+++ frontend/src/shared/index.ts
@@ -1,4 +1,3 @@
-export * from "./components";
-export * from "./containers";
 export * from "./admin-password";
-export * from "./ui/error";
+export * from "./error";
+export * from "./ui";
