+++ frontend/src/shared/ui/components/LoadingState.tsx
