+++ frontend/src/shared/ui/components/index.ts
@@ -1,0 +1,6 @@
+export { default as CenteredContainer } from "./CenteredContainer";
+export { default as LoadingState } from "./LoadingState";
+export { Provider } from "./provider";
+export * from "./color-mode";
+export * from "./toaster";
+export * from "./tooltip";
