--- frontend/src/shared/ui/error/index.ts
@@ -1,4 +1,0 @@
-export { default as ErrorState } from "@/shared/components/error/ErrorState";
-export { default as AlertError } from "@/shared/components/error/AlertError";
-export { default as CardError } from "@/shared/components/error/CardError";
-export { default as CenteredError } from "@/shared/components/error/CenteredError";
